refactor(connectors): keep GitHub commit renderer internal
Move the public-repository commit renderer into a package-internal
module so the coverage seam does not become part of the activity view's
public API.

Strengthen the runtime assertion to verify that multiline commit
messages are truncated and that the rendered commit link keeps its
destination.
